+++ _vercheck.py
@@ -1,5 +1,5 @@
-import aiogram, pydantic, pydantic_core, typing_extensions as te
-print('aiogram', aiogram.__version__)
-print('pydantic', pydantic.version.VERSION)
-print('core', pydantic_core.__version__)
-print('typing-extensions', te.__version__)
+import pydantic, pydantic_core, typing_extensions as te
+
+print("pydantic", pydantic.version.VERSION)
+print("core", pydantic_core.__version__)
+print("typing-extensions", te.__version__)